refactor: move files to appropriate folders
move 'client/src/game/babylon.ts' to 'client/src/game/babylon/babylon.ts'
move 'client/src/game/skybox.ts' to 'client/src/game/skybox/skybox.ts'
update imports accordingly

the purpose of this refactor is to improve the organization and maintainability of the project.
